+++ lib/dialogs/logout-confirmation/index.tsx
@@ -1,121 +1,14 @@
-import React, { Component, Fragment } from 'react';
-import { connect } from 'react-redux';
+import React from 'react';
 
-import AttentionIcon from '../../icons/attention';
-import Dialog from '../../dialog';
-import { closeDialog } from '../../state/ui/actions';
-import { getUnconfirmedChanges } from '../../state/simperium/functions/unconfirmed-changes';
-import exportZipArchive from '../../utils/export';
-import { noteTitleAndPreview } from '../../utils/note-utils';
+import UnsynchronizedConfirmation from '../unsynchronized';
 
-import type * as S from '../../state';
-import type * as T from '../../types';
+const LogoutConfirmation = () => (
+  <UnsynchronizedConfirmation
+    action="REALLY_LOGOUT"
+    actionDescription="Logging out will delete any unsynchronized notes."
+    actionName="Log out"
+    actionSafeName="Safely log out"
+  />
+);
 
-type OwnProps = {
-  action: 'REALLY_CLOSE_WINDOW' | 'REALLY_LOGOUT';
-};
-
-type StateProps = {
-  notes: Map<T.EntityId, T.Note>;
-};
-
-type DispatchProps = {
-  closeDialog: () => any;
-  continueAction: (action: 'REALLY_CLOSE_WINDOW' | 'REALLY_LOGOUT') => any;
-};
-
-type Props = OwnProps & StateProps & DispatchProps;
-
-export class LogoutConfirmation extends Component<Props> {
-  exportUnsyncedNotes = () => {
-    const { closeDialog, notes } = this.props;
-
-    exportZipArchive(notes).then(closeDialog);
-  };
-
-  triggerContinueAction = () => {
-    const { action, continueAction } = this.props;
-    continueAction(action);
-  };
-
-  render() {
-    const { action, closeDialog, notes } = this.props;
-
-    const actionDescription =
-      action === 'REALLY_LOGOUT'
-        ? 'Logging out will delete any unsynchronized notes.'
-        : 'Closing the app with unsynchronized notes could cause data loss.';
-    const actionName = action === 'REALLY_LOGOUT' ? 'Log out' : 'Close window';
-    const actionSafeName =
-      action === 'REALLY_LOGOUT' ? 'Safely log out' : 'Safely close window';
-
-    return (
-      <div className="logoutConfirmation">
-        <Dialog onDone={closeDialog} title="Unsynchronized Notes">
-          <p className="explanation">
-            {notes.size > 0
-              ? actionDescription
-              : 'All notes have synchronized!'}
-          </p>
-
-          {notes.size > 0 && (
-            <Fragment>
-              <p className="explanation-secondary">
-                Possibly unsynchronized notes
-              </p>
-              <section className="change-list">
-                <ul>
-                  {[...notes.entries()].map(([noteId, note]) => {
-                    const { title, preview } = noteTitleAndPreview(note);
-
-                    return (
-                      <li key={noteId}>
-                        <AttentionIcon />
-                        <span className="note-title">{title}</span>
-                      </li>
-                    );
-                  })}
-                </ul>
-              </section>
-            </Fragment>
-          )}
-          {notes.size > 0 && (
-            <button
-              className="export-unsynchronized"
-              onClick={this.exportUnsyncedNotes}
-            >
-              Export unsynchronized notes
-            </button>
-          )}
-
-          <section className="action-button">
-            <button className="log-out" onClick={this.triggerContinueAction}>
-              {notes.size > 0 ? actionName : actionSafeName}
-            </button>
-          </section>
-        </Dialog>
-      </div>
-    );
-  }
-}
-
-const mapStateToProps: S.MapState<StateProps> = (state) => {
-  const changes = getUnconfirmedChanges(state);
-  const notes = new Map(
-    changes.notes.map((noteId: T.EntityId) => [
-      noteId,
-      state.data.notes.get(noteId),
-    ])
-  );
-
-  return {
-    notes,
-  };
-};
-
-const mapDispatchToProps: S.MapDispatch<DispatchProps> = {
-  closeDialog,
-  continueAction: (action) => ({ type: action }),
-};
-
-export default connect(mapStateToProps, mapDispatchToProps)(LogoutConfirmation);
+export default LogoutConfirmation;
