+++ lib/dialogs/unsynchronized/style.scss
