+++ scss/_components.scss
@@ -19,7 +19,7 @@
 @import 'dialogs/import/source-importer/style';
 @import 'dialogs/import/source-importer/executor/style';
 @import 'dialogs/keybindings/style';
-@import 'dialogs/logout-confirmation/style';
+@import 'dialogs/unsynchronized/style';
 @import 'dialogs/settings/style';
 @import 'dialogs/share/style';
 @import 'dialogs/button-group/style';
